logistrack-django/requirements.txt
django==5.2.5
djangorestframework~=3.16.1
drf-yasg~=1.21.10
redis~=6.4.0
mysqlclient~=2.2.7
pydantic~=2.11.7
requests~=2.32.4
django-cors-headers

# testing
pytest~=8.3.2
pytest-django~=4.8.0
pytest-cov~=5.0